requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
chromadb==0.5.5
sentence-transformers==3.1.0
pymupdf==1.25.1
google-genai
python-dotenv==1.0.1
groq
easyocr
numpy
